feign클라이언트 적용, 버그 픽스

+++ src/main/java/com/util/resource/service/CarService.java
@@ -1,8 +1,10 @@
 package com.util.resource.service;
 
+import com.util.dto.SingleResponseDto;
 import com.util.exception.BusinessLogicException;
 import com.util.exception.ExceptionCode;
-import com.util.feign.EmployeeFeignClient;
+import com.util.feign.AuthFeignClient;
+import com.util.feign.dto.EmployeeDto;
 import com.util.resource.entity.Car;
 import com.util.resource.repository.CarRepository;
 import org.springframework.stereotype.Service;
@@ -15,25 +17,24 @@ import java.util.Optional;
 @Service
 public class CarService {
     private final CarRepository carRepository;
-    private final EmployeeFeignClient employeeFeignClient;
+    private final AuthFeignClient authFeignClient;
 
     public CarService(CarRepository carRepository,
-                      EmployeeFeignClient employeeFeignClient) {
+                      AuthFeignClient authFeignClient) {
         this.carRepository = carRepository;
-        this.employeeFeignClient = employeeFeignClient;
+        this.authFeignClient = authFeignClient;
     }
 
     public Car createCar(Car car, long employeeId) throws IllegalArgumentException {
-        // employee 호출할 경우 사용하는 코드
-//        Map<String, Object> employee = employeeFeignClient.getEmployeeById(employeeId);
-//
-//        if (!employee.containsKey("employeeId")) {
-//            throw new BusinessLogicException(ExceptionCode.EMPLOYEE_NOT_FOUND);
-//        }
-//
-//        if (!employee.get("username").equals("관리자")) {
-//            throw new BusinessLogicException(ExceptionCode.CAR_UNAUTHORIZED_ACTION);
-//        }
+        SingleResponseDto<EmployeeDto> employeeDto = authFeignClient.getEmployeeById(employeeId);
+
+        if (employeeDto.getData().getEmployeeId() == null) {
+            throw new BusinessLogicException(ExceptionCode.EMPLOYEE_NOT_FOUND);
+        }
+
+        if (!employeeDto.getData().getName().equals("관리자")) {
+            throw new BusinessLogicException(ExceptionCode.CAR_UNAUTHORIZED_ACTION);
+        }
 
         return carRepository.save(car);
     }
@@ -41,16 +42,15 @@ public class CarService {
     public Car updateCar(Car car, long carId, long employeeId, List<String> imagesToDelete) {
         Car findCar = findVerifiedCar(carId);
 
-        // employee 호출할 경우 사용하는 코드
-//        Map<String, Object> employee = employeeFeignClient.getEmployeeById(employeeId);
-//
-//        if (!employee.containsKey("employeeId")) {
-//            throw new BusinessLogicException(ExceptionCode.EMPLOYEE_NOT_FOUND);
-//        }
-//
-//        if (!employee.get("username").equals("관리자")) {
-//            throw new BusinessLogicException(ExceptionCode.CAR_UNAUTHORIZED_ACTION);
-//        }
+        SingleResponseDto<EmployeeDto> employeeDto = authFeignClient.getEmployeeById(employeeId);
+
+        if (employeeDto.getData().getEmployeeId() == null) {
+            throw new BusinessLogicException(ExceptionCode.EMPLOYEE_NOT_FOUND);
+        }
+
+        if (!employeeDto.getData().getName().equals("관리자")) {
+            throw new BusinessLogicException(ExceptionCode.CAR_UNAUTHORIZED_ACTION);
+        }
 
         Optional.ofNullable(car.getName())
                 .ifPresent(name -> findCar.setName(name));
@@ -85,16 +85,15 @@ public class CarService {
     }
     
     public void deleteCar(long carId, long employeeId) {
-        // employee 호출할 경우 사용하는 코드
-//        Map<String, Object> employee = employeeFeignClient.getEmployeeById(employeeId);
-//
-//        if (!employee.containsKey("employeeId")) {
-//            throw new BusinessLogicException(ExceptionCode.EMPLOYEE_NOT_FOUND);
-//        }
-//
-//        if (!employee.get("username").equals("관리자")) {
-//            throw new BusinessLogicException(ExceptionCode.CAR_UNAUTHORIZED_ACTION);
-//        }
+        SingleResponseDto<EmployeeDto> employeeDto = authFeignClient.getEmployeeById(employeeId);
+
+        if (employeeDto.getData().getEmployeeId() == null) {
+            throw new BusinessLogicException(ExceptionCode.EMPLOYEE_NOT_FOUND);
+        }
+
+        if (!employeeDto.getData().getName().equals("관리자")) {
+            throw new BusinessLogicException(ExceptionCode.CAR_UNAUTHORIZED_ACTION);
+        }
         Car findCar = findVerifiedCar(carId);
 
         carRepository.delete(findCar);
